Isolate change to facilitate moving legacy file
